% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitRHals.R
\name{bootstrap.fitRHals}
\alias{bootstrap.fitRHals}
\title{Bootstrap a fitted Generalized Renshaw-Haberman Model}
\usage{
\method{bootstrap}{fitRHals}(object, nBoot = 1, ...)
}
\arguments{
\item{object}{An \code{fitRHals} object returned by \code{RHalsfitting()}.}

\item{nBoot}{number of bootstrap samples to produce.}
}
\value{
A list with class \code{"bootStMoMo"} with components:

\item{bootParameters}{ a list of of length \code{nBoot} with the fitted 
parameters for each bootstrap replication.} 
\item{model}{ the model fit that has been bootstrapped.}
}
\description{
A generic \code{S3} method that uses residual bootstrap to produce bootstrap  
parameters described in Guo and Li (2024) to account for parameter uncertainty.
}
