% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{Convert_AC_AT}
\alias{Convert_AC_AT}
\title{Convert the Age-Cohort matrix to the Age-Time matrix}
\usage{
Convert_AC_AT(Z)
}
\description{
Convert the Age-Cohort matrix to the Age-Time matrix
}
\keyword{internal}
