% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{Convert_AT_AC}
\alias{Convert_AT_AC}
\title{Convert the Age-Time matrix to the Age-Cohort matrix}
\usage{
Convert_AT_AC(X)
}
\description{
Convert the Age-Time matrix to the Age-Cohort matrix
}
\keyword{internal}
