% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapfitRHals.R
\name{heatmap.fitRHals}
\alias{heatmap.fitRHals}
\title{Draw a Heat-map of Residuals for RH Models}
\usage{
\method{heatmap}{fitRHals}(object)
}
\arguments{
\item{object}{An \code{fitRHals} object returned by \code{RHfit()}.}
}
\value{
This function draws a heat-map (color map) of the residuals using the 
\code{plot()} method from the \code{StMoMo} package.
}
\description{
This function generates a heat-map (color map) of the residuals 
from the fitted model object with class \code{fitRHals} returned by \code{RHfit}. 
The residuals are defined as:
\deqn{d_{x,t} = \log m_{x,t} - \log \hat{m}_{x,t}}
where \eqn{m_{x,t}} are the observed mortality rates, 
and \eqn{\hat{m}_{x,t}} are the fitted values from the model.
Positive residuals are shown in blue, while negative residuals are shown in red.
}
\examples{
# Load example data
Data <- load_EWData(ages = 60:89, years = 1960:2009, series = "Male")

# Fit the RH model using the RHfit function
fit_rh <- RHfit(Data, ages = 60:89, years = 1960:2009)

# Generate the heatmap of residuals
RHheatmap(fit_rh)

}
