% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapfitRHals.R
\name{heatmap}
\alias{heatmap}
\title{Define a generic method for drawing a heat-map of Residuals for RH Models}
\usage{
heatmap(object, ...)
}
\description{
Define a generic method for drawing a heat-map of Residuals for RH Models
}
