% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{iterative_SVD}
\alias{iterative_SVD}
\title{Function for iterative SVD (used for estimating the Renshaw-Haberman model)}
\usage{
iterative_SVD(Z, start.Z_rec = NULL, maxits = 1000, tol = 1e-06)
}
\description{
Function for iterative SVD (used for estimating the Renshaw-Haberman model)
}
\keyword{internal}
