% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{Mx_1x1_EW}
\alias{Mx_1x1_EW}
\title{Mx_1x1_EW}
\format{
A tab-separated text file with columns:
\describe{
  \item{Year}{The year of the data point, ranging from 1841 to 2021.}
  \item{Age}{The age of the individuals, ranging from 0 to 110 and 110+.}
  \item{Female}{The central mortality rate for females.}
  \item{Male}{The central mortality rate for males.}
  \item{Total}{The central mortality rate for the total population.}
}
}
\source{
Data obtained from the Human Mortality Database: \url{http://www.mortality.org/}
}
\description{
A dataset containing mortality rates for England and Wales from 1841 to 2021, including ages from 0 to 110 and 110+.
}
