% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHfit.R
\name{RHfit}
\alias{RHfit}
\title{Fit a Generalized Renshaw-Haberman Model using Least Squares}
\usage{
RHfit(
  data,
  ages,
  years,
  m = 1,
  lc = FALSE,
  const.b0x = FALSE,
  approxConst = FALSE,
  maxits = 10000,
  tol = 1e-06
)
}
\arguments{
\item{data}{A matrix of central mortality rates \eqn{m_{x,t}}, with rows representing ages and columns representing years.}

\item{ages}{A vector of ages corresponding to the rows of the mortality matrix.}

\item{years}{A vector of years corresponding to the columns of the mortality matrix.}

\item{m}{A positive integer corresponding to the number of age-period terms.}

\item{lc}{A logical value controlling whether the cohort effects are not included. If \code{TRUE}, a Lee-Carter model with \eqn{m} age-period terms will be fitted using the order-\eqn{m} singular value decomposition (SVD).}

\item{const.b0x}{A logical value controlling whether \eqn{b_x^{(0)}=1}. If \code{TRUE}, an H1 model will be fitted. Only effective when \code{lc = FALSE}.}

\item{approxConst}{A logical value defining whether the approximate identifiability constraint \eqn{\sum (s-\bar{s})\gamma_s=0} is applied. If \code{TRUE}, a Lagrange multiplier method proposed by Guo and Li (2024) will be implemented. Only effective when \code{lc = FALSE}.}

\item{maxits}{A positive integer corresponding to the maximum number of the main iterations. Only effective when \code{lc = FALSE}.}

\item{tol}{A small positive number corresponding to the tolerance level of the main iterations. Only effective when \code{lc = FALSE}.}
}
\value{
An \code{fitRHals} object containing the following components:
\item{\code{model}}{An \code{StMoMo} object containing the fitted model.}
\item{\code{ax}}{A vector with the fitted values of the static age function \eqn{a_x}.}
\item{\code{bx}}{A matrix with the fitted values of the age-period function \eqn{b_x}.}
\item{\code{kt}}{A matrix with the fitted values of the period indexes \eqn{k_t}.}
\item{\code{b0x}}{A vector with the fitted values of the age-cohort function \eqn{b_x^{0}}.}
\item{\code{gc}}{A vector with the fitted values of the cohort indexes \eqn{\gamma_{t-x}}.}
\item{\code{ssr}}{The sum of squares error of the fitted model.}
\item{\code{npar}}{A positive integer representing the effective number of parameters in the model.}
\item{\code{loglik}}{The Gaussian log-likelihood of the fitted model.}
\item{\code{AIC}}{The Akaike's Information Criterion (AIC) of the fitted model.}
\item{\code{BIC}}{The Bayesian Information Criterion (BIC) of the fitted model.}
\item{\code{ages}}{A vector of ages of the data set.}
\item{\code{years}}{A vector of ages of the data set.}
\item{\code{cohorts}}{A vector of cohorts of the data set.}
\item{\code{mxt}}{The input mortality matrix.}
\item{\code{m}}{The input parameter \eqn{m}.}
\item{\code{lc}}{The input logical variable \code{lc}.}
\item{\code{const.b0x}}{The input logical variable \code{const.b0x}.}
\item{\code{approxConst}}{The input logical variable \code{approxConst}.}
}
\description{
This function fits a RH mortality model using the least squares method.
For the RH model, the fitting is via alternating least squares.
Updating the cohort effects at each iteration uses an iterative SVD.
The model is defined as:
\deqn{
\log m_{x,t} = a_x + \sum_{i=1}^{m}b^{(i)}_x k^{(i)}_t + b_x^{(0)}\gamma_{t-x}
}
}
\examples{
# Example data
Data <- load_EWData(ages = 60:89, years = 1960:2009, series = "Male")
# Fit the RH model using the least squares method
fit_rh <- RHfit(Data, 60:89, 1960:2009)
# Extract the estimated gc
fit_rh$gc
# Plot the parameters estimates using
plot(fit_rh$model)

# Other models

# Fit a Lee-Carter model
fit_lc <- RHfit(Data, 60:89, 1960:2009, lc=T)
# Fit a RH model with two age-period terms
fit_rh2 <- RHfit(Data, 60:89, 1960:2009, m = 2)
# Fit an H1 model
fit_h1 <- RHfit(Data, 60:89, 1960:2009,const.b0x=T)
# Fit an H1 model with the additional identification constraint
fit_h1_appro <- RHfit(Data, 60:89, 1960:2009,const.b0x=T,approxConst=T)
}
