% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHMultifit.R
\name{RHMultifit}
\alias{RHMultifit}
\title{Fit a Multi-Population Generalized Renshaw-Haberman Model using Least Squares}
\usage{
RHMultifit(
  data,
  ages,
  years,
  m = 1,
  lc = FALSE,
  const.b0x = FALSE,
  common.bx = FALSE,
  common.b0x = FALSE,
  maxits = 10000,
  tol = 1e-06
)
}
\arguments{
\item{data}{A list of matrices, where each matrix represents the central mortality rates \eqn{m_{x,t}} for a population, with rows corresponding to ages and columns to years.}

\item{ages}{A vector of ages corresponding to the rows of each mortality matrix.}

\item{years}{A vector of years corresponding to the columns of each mortality matrix.}

\item{m}{A positive integer indicating the number of age-period terms.}

\item{lc}{A logical value controlling whether the cohort effects are excluded. If \code{TRUE}, a Lee-Carter model with \eqn{m} age-period terms will be fitted using the order-\eqn{m} singular value decomposition (SVD).}

\item{const.b0x}{A logical value specifying whether \eqn{b_x^{(0)}=1}. If \code{TRUE}, an H1 model will be fitted. Only effective when \code{lc = FALSE}.}

\item{common.bx}{A logical value indicating whether the age-period parameters \eqn{b_x^{(i)}} should be shared across populations.}

\item{common.b0x}{A logical value indicating whether the age-cohort parameter \eqn{b_x^{(0)}} should be shared across populations. Only effective when \code{lc = FALSE} and \code{const.b0x = FALSE}.}

\item{maxits}{A positive integer representing the maximum number of main iterations. Only effective when \code{lc = FALSE}.}

\item{tol}{A small positive number representing the tolerance level for convergence in the main iterations. Only effective when \code{lc = FALSE}.}
}
\value{
A list containing the following components:
\item{\code{fits}}{A list of \code{fitRHals} objects, one for each population. Each \code{fitRHals} object contains:}
  \itemize{
    \item{\code{model}}{An \code{StMoMo} object containing the fitted model.}
    \item{\code{ax}}{A vector with the fitted values of the static age function \eqn{a_x}.}
    \item{\code{bx}}{A matrix or array with the fitted values of the age-period function \eqn{b_x}.}
    \item{\code{kt}}{A matrix with the fitted values of the period indexes \eqn{k_t}.}
    \item{\code{b0x}}{A vector or array with the fitted values of the age-cohort function \eqn{b_x^{(0)}}.}
    \item{\code{gc}}{A vector with the fitted values of the cohort indexes \eqn{\gamma_{t-x}}.}
    \item{\code{ages}}{A vector of ages of the data set.}
    \item{\code{years}}{A vector of years of the data set.}
    \item{\code{cohorts}}{A vector of cohorts of the data set.}
    \item{\code{mxt}}{The input mortality matrix for the population.}
    \item{\code{m}}{The input parameter \eqn{m}.}
    \item{\code{lc}}{The input logical variable \code{lc}.}
    \item{\code{const.b0x}}{The input logical variable \code{const.b0x}.}
  }
\item{\code{ssr}}{The total sum of squares error across all populations.}
\item{\code{loglik}}{The cumulative Gaussian log-likelihood across all populations.}
\item{\code{npar}}{The effective number of parameters in the multi-population model, adjusted for shared parameters.}
\item{\code{AIC}}{The Akaike's Information Criterion (AIC) for the overall multi-population model.}
\item{\code{BIC}}{The Bayesian Information Criterion (BIC) for the overall multi-population model.}
}
\description{
This function fits a multi-population RH mortality model using the least squares method. It accommodates multiple settings for sharing parameters across populations and allows for different model specifications, including the Lee-Carter, H1, and full RH models.
}
\details{
For the RH and H1 models, the fitting is via alternating least squares with an iterative SVD to update the cohort effects at each iteration.
The general model structure is defined as:
\deqn{
\log m_{x,t} = a_x + \sum_{i=1}^{m}b^{(i)}_x k^{(i)}_t + b_x^{(0)}\gamma_{t-x}
}
}
