% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitRHals.R
\name{seBoot.bootRHals}
\alias{seBoot.bootRHals}
\title{Calculate Standard Errors of Bootstrapped Parameters}
\usage{
\method{seBoot}{bootRHals}(object, ...)
}
\arguments{
\item{object}{An \code{bootRHals} object.}
}
\value{
A list containing the calculated standard error of the parameters:
\code{ax}, \code{bx}, \code{kt}, \code{b0x}, and \code{gc}, each retaining their original structure.
}
\description{
This function calculates the standard error (SE) of the bootstrapped parameters 
from an \code{bootRHals} object.
}
