% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitRHals.R
\name{seBoot}
\alias{seBoot}
\title{Define a generic method for drawing a calculating the standard error
of bootstrapped parameters}
\usage{
seBoot(object, ...)
}
\description{
Define a generic method for drawing a calculating the standard error
of bootstrapped parameters
}
